requirements.txt
streamlit
requests
watchdog
python-dotenv
google-generativeai